# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "description": "Every filter is optional. `period` defaults to `current`, so a call with no query string returns **this budgeting period only**, which is the view TRN-3 draws; pass `period=all` for history. `total` counts matches after filters - read it rather than the array length. There is no pagination: the whole filtered set comes back and the table scrolls.",
        "operationId": "TransactionsController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring of the merchant name. Trimmed; an empty or whitespace-only term filters nothing. **Case-insensitive for ASCII only** - SQLite `LIKE` does not fold non-ASCII, so a merchant name with diacritics matches only on exact case.",
            "schema": {
              "maxLength": 200,
              "example": "groceries",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "One of your categories. An unknown id filters everything out rather than\n404ing - it is a filter, not a resource being addressed.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "One of `current`, `previous`, `all`, or a period `start` in `YYYY-MM-DD` from `GET /api/periods`. Resolved server-side from your pay-schedule history, so the boundary is your budgeting period rather than the calendar month - and periods before a pay-day change keep the boundaries they had. `previous` is the period before the current one. `all` applies no date filter. A date that is not a real calendar date, or that starts none of your periods, is a **400**.",
            "schema": {
              "pattern": "^(current|previous|all|\\d{4}-\\d{2}-\\d{2})$",
              "default": "current",
              "example": "current",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Every sort ends in the same tiebreaks: `date` descending (for the amount sorts only), then `createdAt` descending, then `id`. So the order is stable across requests rather than reshuffling for no visible reason, and two transactions of the same amount read newest-first between themselves.",
            "schema": {
              "default": "date_desc",
              "type": "string",
              "enum": [
                "date_desc",
                "date_asc",
                "amount_desc",
                "amount_asc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List transactions.",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "description": "`amount` is in major units (12.50, not 1250) and must be positive. `date` is a calendar date, `YYYY-MM-DD`, stored verbatim - backdating is supported and puts the transaction in the month that date falls in. **404** means the `categoryId` in the body names no category of yours.",
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a transaction.",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/transactions/scan": {
      "post": {
        "description": "Every field is null when it could not be read or failed validation against live data, with `missing` naming which of `merchant`, `amount`, `date` and `categoryId` came back that way. **503** means scanning is not configured (no `GEMINI_API_KEY`), and **504** means the extraction call timed out - both distinct from a 200 with everything missing, which means the photo was read but nothing on it was legible.",
        "operationId": "TransactionsController_scan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "At most 4 images (pages of one receipt), or exactly one PDF."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanReceiptResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "The upload exceeded its size cap. `message` names which cap was passed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "The feature is not configured on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "A downstream call did not finish in time. Safe to retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Extract merchant, amount, date, category and note from a receipt.",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "description": "Carries two extras on **different windows**, deliberately. `category` is that category's progress for the **current** period, even when this transaction is from an earlier one - the bar answers where the category stands now. `recentInCategory` is the latest transactions in the same category from **any** month, excluding this one. **404** here always means the id in the URL. Expect `category.status` to be `uncapped` most of the time: caps are optional and the preselected fallback has none.",
        "operationId": "TransactionsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One transaction, with the context frame 08 draws around it.",
        "tags": [
          "transactions"
        ]
      },
      "patch": {
        "description": "Send only the fields to change: an absent field is left alone, and `note` accepts null to clear it. An empty body is a **400**, because it would record an edit that changed nothing. **404** means either the id in the URL or a `categoryId` in the body names nothing of yours.",
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a transaction.",
        "tags": [
          "transactions"
        ]
      },
      "delete": {
        "description": "Permanent as far as this API is concerned: the transaction stops existing through every endpoint. Deleting one twice is a **404**, and **404** here always means the id in the URL.",
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted. No body."
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a transaction.",
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Spendifico API",
    "description": "The HTTP contract of the Spendifico backend.\n\nGenerated from the NestJS code at build time, committed as `backend/openapi.json`, and\nthe only source of the frontend response types in `frontend/src/types/api.d.ts`.\nDo not edit either by hand: run `npm run api:sync` from the repo root.\n\nEvery operation can additionally answer **500** with `ErrorResponseDto`, via the global\nexception filter. It is stated once here rather than per operation: it is the same\nnon-actionable fact everywhere, and repeating it would widen every generated response type.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "The raw session token returned by POST /api/auth/verify. Opaque: 256 random bits, base64url, looked up by hash server-side."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "description": "String for most errors; an array for class-validator's field messages.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "statusCode": {
            "type": "number"
          },
          "error": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "statusCode",
          "error",
          "timestamp",
          "path"
        ]
      },
...
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "merchant": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "description": "Major units (e.g. 12.5). Stored as integer cents; converted on the way out."
          },
          "date": {
            "type": "string",
            "description": "`YYYY-MM-DD`, exactly the string that was stored."
          },
          "note": {
            "type": "string",
            "nullable": true,
            "description": "Null when the transaction has no note, never absent."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601."
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601. Within a millisecond of `createdAt` until the first edit, not\nnecessarily equal to it: the two columns default from independent `new Date()`\ncalls, so an insert can straddle a millisecond boundary. Do not use equality\nof the two to mean \"never edited\"."
          }
        },
        "required": [
          "id",
          "merchant",
          "categoryId",
          "amount",
          "date",
          "note",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "TransactionsResponseDto": {
        "type": "object",
        "properties": {
          "transactions": {
            "description": "Every matching transaction, in the requested sort order.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Matches after every filter, not the account total. Equal to `transactions.length` while there is no pagination; read this rather than the array length, so a future page size cannot silently turn the badge into a page count."
          },
          "period": {
            "nullable": true,
            "description": "The period this list covers, echoing back what `?period=` resolved to - use `label` for the screen’s overline. **Null when `period=all`**, which spans every period and so has no single label.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PeriodSummaryDto"
              }
            ]
          }
        },
        "required": [
          "transactions",
          "total",
          "period"
        ]
      },
      "ScanReceiptResponseDto": {
        "type": "object",
        "properties": {
          "merchant": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "categoryId": {
            "type": "string",
            "nullable": true,
            "description": "An id from the caller's own live categories, verbatim, or null. Never a hallucinated id: `ReceiptScanService` validates it against the same list the prompt was given.",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "missing": {
            "type": "array",
            "description": "Which of `merchant`, `amount`, `date` and `categoryId` came back null.",
            "items": {
              "type": "string",
              "enum": [
                "merchant",
                "amount",
                "date",
                "categoryId"
              ]
            }
          }
        },
        "required": [
          "merchant",
          "amount",
          "date",
          "categoryId",
          "note",
          "missing"
        ]
      },
...
      "TransactionDetailResponseDto": {
        "type": "object",
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TransactionResponseDto"
          },
          "category": {
            "description": "The transaction's category with its stats for the **current** period - not the period the transaction itself falls in. The bar answers where this category stands now.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryResponseDto"
              }
            ]
          },
          "recentInCategory": {
            "description": "Up to 5 other transactions in the same category, newest first, from any month. Excludes the one in `transaction`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponseDto"
            }
          }
        },
        "required": [
          "transaction",
          "category",
          "recentInCategory"
        ]
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Major units (e.g. 12.50), stored as integer cents. Positive: this app\nrecords spending, and direction is not a per-row choice.",
            "maximum": 1000000000,
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "date": {
            "type": "string",
            "description": "The calendar day the money was spent, `YYYY-MM-DD`. Stored verbatim.\n\nBackdating is ordinary and supported: the month a transaction belongs to is\nderived from this at read time, so a date in a past month lands in that\nmonth rather than the one it was entered in.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "format": "date",
            "example": "2026-08-03"
          },
          "merchant": {
            "type": "string",
            "maxLength": 200
          },
          "categoryId": {
            "type": "string",
            "description": "An existing category of this user's. An unknown id is a 404, not a 400.",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "amount",
          "date",
          "merchant",
          "categoryId"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Major units, as on create.",
            "maximum": 1000000000,
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "date": {
            "type": "string",
            "description": "`YYYY-MM-DD`, as on create. Inline regex for the `pattern` lift.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "format": "date",
            "example": "2026-08-03"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "description": "The one nullable field, and so the one that keeps `@IsOptional()`: null is\na meaningful value here rather than a mistake, and it means \"clear the\nnote\".",
            "maxLength": 500
          },
          "merchant": {
            "type": "string",
            "maxLength": 200
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}